refactored main grid into components and added divider
