+++ client/components/weigh-sources.js
